by popular demand: a makefile that is truly GNUmake and BSDmake compatible 	(the explanation below is NOT for the faint of heart)
why was this non-trivial?
BSDmake implcitly includes ".depend" if it exists, for convenience of
	throwing all your results from gcc -MM into a dependency file
GNUmake completely ignores it, and thus requires explicit inclusion
However, any file that BSD explicitly includes and doesn't exist,
	will cause it to die, EVEN if it's .depend
	(whereas gmake is happy to ignore non-existent includes)
This is further complicated by the fact that conditional directives and
	expansions are entirely incompatible between the two make
	systems (".if" vs. "if", etc...)
So we have a chicken-before-the-egg/cart-before-the-horse problem.
The Makefile must not explicitly include .depend until it exists.
Thus, we SELF-MODIFY the makefile to append "include .depend"
	after it's been built, and remove that line during a
	"make clobberdepend" (also in make commit)
	This safely maintains functionality of the .depend file in
	both make systems without causing either to fail.
	Caveat: This will fail if the end-user (on BSDmake) removes the
	.depend file by a plain "rm".  This is not a common thing
	for a user to do, so we accept this Achilles-heel.

the moral of this story:
	BSDmake sucks.  alias make gmake.