The not-so-long awaited:
bison-output-to-C.awk, which complements the yacc version.

Converts part of parser state machine description (y.output) into string
that can be used in C (y.output.h), particularly useful in reporting
syntax errors precisely, and verbosely.  Whoop-de-doo.

However, do NOT expect the regression tests to report the same
output as those used with the yacc version because bison performs
some state-machine optimization resulting in more efficient state
transitions (e.g. merging chains of goto's).

May eventually think of a way to resolve this, but is not important
for the moment.

The Makefiles have not been changed to use this yet.
Eventually need autoconf to figure this out...