The long awaited...
GNU-ART++ 0.1.1 release.

Ok is not really named GNU ART, but the only accomplishment in this release
is porting the entire project to use GNU's build system: autoconf (2.59),
automake (1.9), and libtool (1.5.6).  (and bison (1.875), too!)

Configuration files and scripts that are automatically generated are also
included in the CVS repository, so that developers are not required
to use the same tools, but nevertheless, they are highly recommended.
To purge them, run the "maintainerclobber" script in the distclean state,
and then run the "bootstrap" script to recreate them.

The source files have been reorganized into subdirectories,
and the lexer and parser have been re-done.
In particular the parser input is now compatible with both bison and yacc,
with a few underhanded awk tricks to achieve this feat.
<Insert long-winded description on the incompatibilities and interface
changes from yacc to bison.>

Configuration properly detects bison vs. yacc, the presence of
doxygen and dot (only required for documentation generation).

The test suite has been revamped to leverage the built-in testing
features from automake.
Instead of flat operation, tests are run recursively by directories.
However, as soon as one test fails in a directory, the entire test
will terminate.

And finally, make distcheck (which uses a VPATH build) passes,
which took two days to get working.

blahbety blahbety blah...

Details: trust me, you don't want to know.

Now we can return to productive coding...