branch:	ARTXX-00-01-04-main-00-04-object-01-03-unroll-00 rev:	ARTXX-00-01-04-main-00-04-object-01-03-unroll-00-01
This snapshot is -UNSTABLE-.

Everything will compile but all tests will fail because debug code is on.
The goal of this branch is to get template/process in the test directory
working.

We will next attempt to merge (on my knees praying) revisions from
the -persistent-00-03 sub-series.
These revisions are needed to get instance collections working...