branch:	ARTXX-00-01-04-main-00-15-collect rev:	ARTXX-00-01-04-main-00-15-collect-02
Upgraded persistent_object_manager's reconstruction_table
to util::list_vector<reconstruction_table_entry>.
NEVERMORE will the entire array of entries be reallocated to
new locations upon push_back()!
Accesses to the array are now O(log(N/C)) where N is the
total size, and C is the chunk-size (because of internal map).

That wasn't so bad...
This branch is ready to merge to parent.