branch:	ARTXX-00-01-04-main-00-36-PRS-01 rev:	ARTXX-00-01-04-main-00-36-PRS-01-07
This revision doesn't really belong here, but it was done anyhow.
What's good for the gander, is good for the goose.
Most changes apply to the util library, almost nothing is specific
	to art++ itself.
Most notable changes include improvement of count_ptr to use pools of
	pre-allocated counts, instead of new size_t each time.
See src/ChangeLog for more details.