fixes ACX-PR-10053: failure to propagate some attributes on process/struct ports up the hierarchy. This is related to the way port aliases are replayed.
Internal changes:

During replay of port aliases, attributes and actuals of aliased ports are
*assumed* to already match (as enforced when that type was created),
so now, just reconnect aliases' next pointers, and assert that actuals
and attributes match.

On instance-alias instantiate(), replay port aliases as soon as type
is complete.  There may still be a pass that does this redundantly,
but will clean that up later.

In export_alias_properties(), apply properties to all aliases in each set,
not just the canonical (top) one.